Add rule to detect surprising use of version with package upgrade
Per [this issue](https://github.com/chef/chef-web-docs/pull/45), using the `package` or `yum_package` resource with the `:upgrade` action and specifying a version number can lead to unexpected behavior.

When `:upgrade` is used, the `version` attribute is silently ignored and the package is upgraded to the latest version available.  One might have expected it to upgrade to the desired version.